User photo and name now displayed when signed in.
The user name and photo are now displayed when the user is signed in. I
do not currently check if the user does have a profile image. It is also
worth noting that when a new user signs up, there is no user information
storage created in the Firestore. This should be implemented by the
Flask backend server.